src/web/backend/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Tina Huang <tina@monkey.name>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chalice = "^1.31.0"
requests = "^2.31.0"
boto3 = "^1.34.104"
chalice-graphql = "^0.1.0"
graphql-core = "^3.2.3"
ariadne = "^0.23.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
